Add Prominent page
Making the prominent page with css is kind of hard. Ideally it should consist of
buttons that each have the same width and height.
However when a board has 5 members instead of 4 then that board should still
take the same amount of space as a different board.

Additionally we should have both space for board members and other prominent
members, so we have to find a way to divide that space.