Add simple authentication
Currently we have a simple form that can be used to send a password to the
server from which we will get a jwt token that can be used for all further
requests.

We should add some feedback when sending the authentication request.
Additionally there should be a redirect if the current user is not authenticated.