Change test utils to not use memory router
This make the application we test better reflect thereal application.

I'm not yet sure if this has performance consequences since the
MemoryRouter was specifically designed to be used as a testing utility.
However it is difficult to use the memory router with the
connected-react-rotuer which resulted in us needing to use
flushAllPromises in the past.
Therefore the goal is to use the normal router for now, which works with
connected-react-router and reflects the real life usage.
Later once we remove connected-react-router we might be able to move
back to the memory router.