Extract authentication from redux
The AuthenticationProvider has been refactored so that all of its state
is kept locally, instead of using the global redux store.
This makes it simpler, allowing us more easily test it and use it
outside of the context of a redux store.

Currently the AuthenticationForm still relies on dispatch so that we
fetch initial data after authenticating.
This will be removed after we add a new context containing this initial data.