Get products from ProductsContext using react-query
Moving to react-query will make our async actions much simpler, not
requiring the _request, _success and _failure actions.

To do this we will have to change the loading screen so that it uses the
new products query instead of the loading features reducer, which will
result in a bit more code right now, but that will be fixed once we
refactor the loading screen.