Get current queued order using maxBy
The previous method was hacky: it sorted the orders and assumed a
"first" order always existed.

Using the maxBy method is much better as it more clearly shows the
intent and doesn't need to take an element from an array.

An advantage of this better approach is that it is potentially more
performant, though that's a nice side effect and isn't relevant at all
for this app.