Store queued orders in local storage
We use a modified useLocalStorage hook to store the queued orders. The
difference being that we only need it to save and load the orders, not
to sync them across different browser tabs, though this would be a nice
advantage.

The timeout functions makes dealing with this a bit messy since their
closures capture the current state.
I think we should first refactor those timeouts and then reconsider the
local storage implementation.