Remove AppContainer used only by tests
We now reuse the same app routes in app-container.test.tsx and app.test.tsx
These tests still aren't that great, i.e. when they fail it takes quite a bit of
effort to make them pass. But they do give confidence that there is no bug in
the code so I will keep them as is for now.